Use notification bar for updates
Replaces the Yes/No dialog at the start of the application with a less intrusive notification bar at the top.
When the current version is too old or the architecture is wrong, this bar is not dismissable,
additionally, connecting to a Wi-Fi network is blocked by disabling the "Connect to eduroam" button,
and by disabling the "Next" button when viewing a profile, the latter to prevent connecting when opening an EAP-config file.

An optional dropdown menu besides the Update button is available, to visit a website where the program can be obtained. When the application is running on the wrong architecture and no update to the same or newer version is available for the correct architecture, this button will always send the user to this website.